chatbot/requirements.txt
fastapi==0.115.12
langchain==0.3.23
langchain_community==0.3.21
langchain_openai==0.3.12
pydantic==2.11.3
python-dotenv==1.1.0
uvicorn==0.34.0
